chore: デコレータを移動
